Order as parameter of Rest services for saving picture. It has the string representing the picture

+++ src/main/java/fr/xebia/photobooth/api/PhotoResource.java
@@ -4,12 +4,12 @@ import fr.xebia.photobooth.domain.Order;
 import fr.xebia.photobooth.domain.Validation;
 
 import javax.ws.rs.*;
-import javax.ws.rs.Path;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.io.*;
 import java.net.URL;
-import java.nio.file.*;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.Base64;
 
 
@@ -31,18 +31,24 @@ public class PhotoResource {
 
 
     @POST
-    @Path("/check")
+    @Path("/validate")
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response checkCommand(Order order) {
+    public Response validateCommand(Order order) {
         return Response.ok().entity(Validation.INSTANCE.isValid(order)).build();
     }
 
     @POST
     @Path("/save")
     @Consumes(MediaType.APPLICATION_JSON)
-    public String saveToFile(String base64Picture) throws IOException {
+    public String saveToFile(Order order) throws IOException {
+        if(!Validation.INSTANCE.isValid(order)){
+            throw new WebApplicationException(Response.status(Response.Status.FORBIDDEN)
+                    .entity("Not a valid order")
+                    .build());
+        }
+
         File urlFile = File.createTempFile("image", ".png");
-        byte[] data = Base64.getDecoder().decode(base64Picture);
+        byte[] data = Base64.getDecoder().decode(order.stringFile);
 
         try (OutputStream stream = new FileOutputStream(urlFile);) {
             stream.write(data);
@@ -57,23 +63,23 @@ public class PhotoResource {
     @POST
     @Path("/saveWithURL")
     @Consumes(MediaType.APPLICATION_JSON)
-    public String saveToFileWithURL(String pictureUrl) throws IOException {
+    public String saveToFileWithURL(Order order) throws IOException {
+        if(!Validation.INSTANCE.isValid(order)){
+            throw new WebApplicationException(Response.status(Response.Status.FORBIDDEN)
+                    .entity("Not a valid order")
+                    .build());
+        }
+
         File urlFile = File.createTempFile("image", ".png");
         urlFile.delete();
 
-        try (InputStream in = new URL(pictureUrl).openStream()) {
+        try (InputStream in = new URL(order.stringFile).openStream()) {
             Files.copy(in, urlFile.toPath());
         }
-
         java.nio.file.Path source = urlFile.toPath();
         Files.move(source, Paths.get("src/main/webapp").resolve(source.getFileName()));
 
+        //return Response.ok().entity(urlFile.getName()).build();
         return urlFile.getName();
     }
-
-	
-	/*
-     *
-	 * // Note preferred way of declaring an array variable
-	 */
 }
